Move model integration test to tests folder
